+++ ts/common/versionInfo/saga/backendInfo.ts
@@ -3,22 +3,22 @@ import * as t from "io-ts";
 import { readableReport } from "italia-ts-commons/lib/reporters";
 import { BasicResponseType } from "italia-ts-commons/lib/requests";
 import { call, Effect, fork, put } from "redux-saga/effects";
-import { VersionInfo } from "../../definitions/content/VersionInfo";
-import { ContentClient } from "../api/content";
+import { VersionInfo } from "../../../../definitions/content/VersionInfo";
+import { ContentClient } from "../../../api/content";
 import {
   backendInfoLoadFailure,
   backendInfoLoadSuccess
-} from "../store/actions/backendInfo";
-import { SagaCallReturnType } from "../types/utils";
-import { startTimer } from "../utils/timer";
+} from "../../../store/actions/backendInfo";
+import { SagaCallReturnType } from "../../../types/utils";
+import { startTimer } from "../../../utils/timer";
 
-// load backend info every hour
-const BACKEND_INFO_LOAD_INTERVAL = 60 * 60 * 1000;
+// load version info every hour
+const VERSION_INFO_LOAD_INTERVAL = 60 * 60 * 1000;
 
-// retry loading backend info every 10 seconds on error
-const BACKEND_INFO_RETRY_INTERVAL = 10 * 1000;
+// retry loading version info every 10 seconds on error
+const VERSION_INFO_RETRY_INTERVAL = 10 * 1000;
 
-function* backendInfoWatcher(): Generator<Effect, void, any> {
+function* versionInfoWatcher(): Generator<Effect, void, any> {
   const contentClient = ContentClient();
 
   function getServerInfo(): Promise<
@@ -33,23 +33,23 @@ function* backendInfoWatcher(): Generator<Effect, void, any> {
 
   while (true) {
     try {
-      const backendInfoResponse: SagaCallReturnType<typeof getServerInfo> =
+      const versionInfoResponse: SagaCallReturnType<typeof getServerInfo> =
         yield call(getServerInfo);
       if (
-        backendInfoResponse.isRight() &&
-        backendInfoResponse.value.status === 200
+        versionInfoResponse.isRight() &&
+        versionInfoResponse.value.status === 200
       ) {
-        yield put(backendInfoLoadSuccess(backendInfoResponse.value.value));
-        yield call(startTimer, BACKEND_INFO_LOAD_INTERVAL);
+        yield put(backendInfoLoadSuccess(versionInfoResponse.value.value));
+        yield call(startTimer, VERSION_INFO_LOAD_INTERVAL);
       } else {
-        const errorDescription = backendInfoResponse.fold(
+        const errorDescription = versionInfoResponse.fold(
           readableReport,
           ({ status }) => `response status ${status}`
         );
 
         yield put(backendInfoLoadFailure(new Error(errorDescription)));
 
-        yield call(startTimer, BACKEND_INFO_RETRY_INTERVAL);
+        yield call(startTimer, VERSION_INFO_RETRY_INTERVAL);
       }
     } catch (e) {
       yield put(backendInfoLoadFailure(new Error(e)));
@@ -58,5 +58,5 @@ function* backendInfoWatcher(): Generator<Effect, void, any> {
 }
 
 export default function* root(): IterableIterator<Effect> {
-  yield fork(backendInfoWatcher);
+  yield fork(versionInfoWatcher);
 }
